requirements.txt
aniso8601==10.0.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==2.3.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
greenlet==3.2.1
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.3.2
MarkupSafe==3.0.2
numpy==1.24.3
packaging==25.0
pandas==2.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.17.0
SQLAlchemy==2.0.40
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==2.3.4
